you can no longer delete question from test when there is only one left and also the questions numbers are changed after a question is deleted

+++ TestingEz/src/main/java/com/testingez/testingez/services/impls/QuestionServiceImpl.java
@@ -20,10 +20,12 @@ import com.testingez.testingez.repositories.TestRepository;
 import com.testingez.testingez.services.QuestionService;
 import com.testingez.testingez.services.ResultService;
 import com.testingez.testingez.services.UserHelperService;
+import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import java.util.Comparator;
 import java.util.List;
 
 @AllArgsConstructor
@@ -95,8 +97,8 @@ public class QuestionServiceImpl implements QuestionService {
             return null;
         }
         Question question = this.questionRepository.findByTestIdAndNumber(testId, questionNumber)
-                .orElseThrow(() -> new QuestionNotFoundException("Question(№" + questionNumber + ") associated with test: " + testId +
-                        "was not found!"));
+                .orElseThrow(() -> new QuestionNotFoundException("Question №" + questionNumber + " associated with test: " + testId +
+                        " was not found!"));
         QuestionAnswerDTO map = this.modelMapper.map(question, QuestionAnswerDTO.class);
         map.setResponseTime(test.getResponseTime());
         map.setTestId(testId);
@@ -226,14 +228,32 @@ public class QuestionServiceImpl implements QuestionService {
      */
     @Override
     public Boolean delete(Long questionId, Long testId) {
+        // TODO: touch the numbering when a question is deleted
         if (this.resultRepository.countByTestId(testId) > 0) {
             return false;
         }
         Test test = this.testRepository.findById(testId)
                 .orElseThrow(() -> new TestNotFoundException("We couldn't find test with id: " + testId));
+        if (test.getQuestionsCount() <= 1) {
+            return false;
+        }
         test.setQuestionsCount(test.getQuestionsCount() - 1);
         this.questionRepository.deleteById(questionId);
+        touchNumbering(testId);
         return true;
+    }
+
+    private void touchNumbering(Long testId) {
+        List<Question> questions = this.questionRepository.findAllByTestId(testId)
+                .stream().sorted(Comparator.comparing(Question::getNumber)).toList();
+        int n = questions.size();
+        for (int i = 1; i <= n; i++) {
+            Question question = questions.get(i - 1);
+            if (question.getNumber() != i) {
+                question.setNumber(i);
+                this.questionRepository.saveAndFlush(question);
+            }
+        }
     }
 
 }
